nohup/python/generate-oficios/pyproject.toml
[project]
name = "python-shell-pdf"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "weasyprint",
    "python-dotenv",
    "jinja2",
    "fonttools>=4.60.1",
]
